move lecture and search types to types directory
