+++ web-ui/src/components/image-search/image-search.tsx
@@ -31,7 +31,7 @@ import {
 import { UploadChangeParam, UploadFile } from 'antd/es/upload';
 import { Image as ImageType, Question } from '../search';
 import QuestionInput from '../question-input/question-input';
-import { Lecture } from '../lecture';
+import { Lecture } from '../../types/lecture';
 import { PreviewCompact } from '@/components/lecture/preview';
 import { history } from 'umi';
 import { QuestionAnswer } from './question-answer';
