+++ web-ui/src/components/tables/denied-table.tsx
@@ -3,7 +3,7 @@ import { notification } from 'antd';
 import { useEffect, useState } from 'react';
 import { useMutation } from '@tanstack/react-query';
 import apiClient, { ServerErrorResponse, ServerResponse } from '@/http';
-import { Lecture } from '@/components/lecture';
+import { Lecture } from '@/types/lecture';
 import styles from './denied-table.less';
 import { ColumnsType } from 'antd/es/table';
 import svFlag from '@/assets/flag-sv.svg';
