+++ web-ui/src/types/lecture.ts
